# Conflicts:
#	src/Interpreters/Cache/FileCacheSettings.cpp

+++ src/Common/examples/thread_creation_latency.cpp
@@ -5,6 +5,7 @@
 
 #include <Common/Stopwatch.h>
 #include <Common/Exception.h>
+#include <Common/ErrnoException.h>
 #include <Common/ThreadPool.h>
 #include <Common/CurrentMetrics.h>
 
@@ -60,8 +61,8 @@ void test(size_t n, const char * name, F && kernel)
         << std::fixed << std::setprecision(2)
         << n << " ops in "
         << watch.elapsedSeconds() << " sec., "
-        << n / watch.elapsedSeconds() << " ops/sec., "
-        << "avg latency: " << watch.elapsedSeconds() / n * 1000000 << " μs, "
+        << static_cast<double>(n) / watch.elapsedSeconds() << " ops/sec., "
+        << "avg latency: " << watch.elapsedSeconds() / static_cast<double>(n) * 1000000 << " μs, "
         << "max latency: " << max_seconds * 1000000 << " μs "
         << "(res = " << value << ")"
         << std::endl;
